Fix #71: Modularize tree.rs into submodules
Split 966-line tree.rs into focused submodules:
- tree/config.rs - WalkerConfig type
- tree/filter.rs - FileFilter enum
- tree/json_types.rs - JSON serialization types
- tree/walker.rs - TreeWalker implementation
- tree/streaming.rs - StreamingWalker implementation
- tree/utils.rs - Shared utilities (glob matching, file size formatting)
